requirements.txt
Flask
Flask-CORS
PyJWT
PyPDF2
google-generativeai
python-dotenv
requests
Werkzeug
